Move log functions to own module
