+++ src/main.js
@@ -1,7 +1,8 @@
 import mqtt from "mqtt";
 import QRCode from "qrcode";
 import { version } from "../package.json";
-import { settings, initSettingsTabs, loadSettings } from "./settings";
+import { initSettingsTabs, loadSettings } from "./settings";
+import { createPlainLogEntry, createUsercodeLogEntry, initLog } from "./logs.mjs";
 
 const options = {
   keepalive: 30,
@@ -15,15 +16,12 @@ const options = {
 
 const brokerHost = localStorage.getItem("brokerHost") ?? location.hostname;
 const client = mqtt.connect(`ws://${brokerHost}:9001`, options);
-const logMessageRegex = /\[(\d+:\d{2}:\d{2}\.?\d*)] (.*)/;
 let connectedServices = {
   astdiskd: false,
   astmetad: false,
   astprocd: false,
 };
 let $ = {};
-let shouldAutoScroll = true;
-let generatedScrollEvent = false;
 
 function updateServiceState() {
   const runningServiceCount = Object.values(connectedServices).filter(
@@ -39,11 +37,7 @@ function updateServiceState() {
 
 window.addEventListener("DOMContentLoaded", (event) => {
   $ = {
-    log: document.getElementById("log"),
     wifiQRCode: document.getElementById("qrcode-wifi"),
-    templates: {
-      logEntry: document.getElementById("tpl-log-entry"),
-    },
     modals: {
       disconnected: document.getElementById("modal-disconnected"),
       info: document.getElementById("modal-info"),
@@ -52,39 +46,15 @@ window.addEventListener("DOMContentLoaded", (event) => {
     noAnnotatedImageInstructions: document.getElementById(
       "no-annotated-image-instructions",
     ),
-    scrollToBottom: document.getElementById("scroll-to-bottom"),
   };
 
   initSettingsTabs();
   loadSettings();
-
-  /// Autoscroll
-  $.log.addEventListener(
-    "scroll",
-    function (e) {
-      if (generatedScrollEvent) {
-        generatedScrollEvent = false;
-      } else {
-        shouldAutoScroll = false;
-        $.log.dataset.autoscroll = "false";
-      }
-    },
-    {
-      passive: true,
-    },
-  );
-
-  $.scrollToBottom.addEventListener("click", function (e) {
-    shouldAutoScroll = true;
-    generatedScrollEvent = true;
-    $.log.dataset.autoscroll = "true";
-    $.log.scrollTop = $.log.scrollHeight;
-  });
+  initLog();
 
   document.getElementById("info-kit-ui-version").textContent = version;
 
   /// Modals
-
   // Add a click event on modal triggers
   document.querySelectorAll(".modal-trigger").forEach(($trigger) => {
     const modal = $trigger.dataset.target;
@@ -198,41 +168,8 @@ client.on("error", function (err) {
 
 client.on("close", disconnected);
 
-function checkScrollbackLimit() {
-  if (settings.scrollbackLimit.value !== -1 && $.log.childElementCount > settings.scrollbackLimit.value) {
-    $.log.firstElementChild.remove();
-  }
-}
-
 const handlers = {
-  "astoria/broadcast/usercode_log": (contents) => {
-    checkScrollbackLimit();
-
-    const template = $.templates.logEntry;
-    const entryFragment = template.content.cloneNode(true);
-    const [_, ts, message] = contents.content.match(logMessageRegex);
-
-    entryFragment.querySelector(".log-entry").dataset.source = contents.source;
-    entryFragment.querySelector(".log-entry__ts").textContent = ts;
-    const contentEl = entryFragment.querySelector(".log-entry__content");
-    contentEl.innerText = message.replaceAll(" ", String.fromCharCode(0xa0));
-
-    if (contents.source === "astoria") {
-      contentEl.classList.add(
-        "has-text-weight-bold",
-        "has-text-centered",
-        "is-family-sans-serif",
-      );
-    } else if (contents.source === "stderr") {
-      contentEl.classList.add("has-text-danger");
-    } else if (message.indexOf("WARNING:") === 0) {
-      contentEl.classList.add("has-text-warning");
-    }
-
-    $.log.appendChild(entryFragment);
-    generatedScrollEvent = true;
-    if (shouldAutoScroll) contentEl.scrollIntoView({ block: "end" });
-  },
+  "astoria/broadcast/usercode_log": createUsercodeLogEntry,
   "astoria/broadcast/start_button": (contents) => {
     createPlainLogEntry(
       "Start button pressed",
@@ -325,30 +262,6 @@ function uuid4() {
       v = c === "x" ? r : (r & 0x3) | 0x8;
     return v.toString(16);
   });
-}
-
-function createPlainLogEntry(text, icon = null, icon_class = null, ...classes) {
-  checkScrollbackLimit();
-  const entry = document.createElement("div");
-  entry.classList.add("plain-log-entry", ...classes);
-  entry.textContent = text;
-
-  if (icon) {
-    const iconElement = document.createElement("span");
-    iconElement.classList.add("material-symbols-outlined");
-    icon_class && iconElement.classList.add(icon_class);
-    iconElement.textContent = icon;
-    entry.prepend(iconElement);
-  }
-
-  $.log.appendChild(entry);
-
-  if (shouldAutoScroll) {
-    generatedScrollEvent = true;
-    entry.scrollIntoView();
-  }
-
-  return entry;
 }
 
 function sendProcessRequest(type) {
